+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/note/types/PodcastEpisode.kt
@@ -22,10 +22,8 @@ package com.vitorpamplona.amethyst.ui.note.types
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
@@ -33,7 +31,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
@@ -42,22 +39,12 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.vitorpamplona.amethyst.commons.model.toImmutableListOfLists
 import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.service.playback.composable.GetVideoController
-import com.vitorpamplona.amethyst.service.playback.composable.PauseControllerWhenInBackground
-import com.vitorpamplona.amethyst.service.playback.composable.WaveformData
-import com.vitorpamplona.amethyst.service.playback.composable.mediaitem.GetMediaItem
 import com.vitorpamplona.amethyst.ui.components.TranslatableRichTextViewer
 import com.vitorpamplona.amethyst.ui.navigation.navs.INav
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.AccountViewModel
 import com.vitorpamplona.amethyst.ui.theme.Size5dp
 import com.vitorpamplona.amethyst.ui.theme.replyModifier
 import com.vitorpamplona.quartz.nipF4Podcasts.episode.PodcastEpisodeEvent
-
-// NIP-F4 doesn't carry a waveform tag, so every episode renders with the same flat baseline.
-// Hoist to a top-level constant so we share one List<Float>+WaveformData across the whole
-// feed instead of allocating a fresh 96-element list per visible card.
-private const val WAVEFORM_SAMPLES = 96
-private val FLAT_WAVEFORM = WaveformData(List(WAVEFORM_SAMPLES) { 0.4f })
 
 // Bottom-rounded border on the audio player so it visually butts up against the cover's
 // top-rounded corners as one card. Constant — keep out of recomposition.
@@ -88,39 +75,14 @@ fun RenderPodcastEpisode(
     Column(MaterialTheme.colorScheme.replyModifier) {
         PodcastCoverCard(image, note, accountViewModel)
         firstAudio?.let { audio ->
-            val callbackUri = remember(noteEvent) { note.toNostrUri() }
-
-            Row(
-                Modifier.fillMaxWidth().height(80.dp),
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                GetMediaItem(
-                    videoUri = audio.url,
-                    title = title,
-                    artworkUri = image,
-                    authorName = note.author?.toBestDisplayName(),
-                    callbackUri = callbackUri,
-                    mimeType = audio.mediaType,
-                    aspectRatio = null,
-                    proxyPort = accountViewModel.httpClientBuilder.proxyPortForVideo(audio.url),
-                    keepPlaying = false,
-                    waveformData = FLAT_WAVEFORM,
-                ) { mediaItem ->
-                    GetVideoController(
-                        mediaItem = mediaItem,
-                        muted = false,
-                    ) { controller ->
-                        PauseControllerWhenInBackground(controller)
-                        RenderVoicePlayer(
-                            mediaItem = mediaItem,
-                            controllerState = controller,
-                            waveform = FLAT_WAVEFORM,
-                            borderModifier = PLAYER_BORDER_MODIFIER,
-                            accountViewModel = accountViewModel,
-                        )
-                    }
-                }
-            }
+            PodcastEpisodeAudioPlayer(
+                audio = audio,
+                note = note,
+                title = title,
+                image = image,
+                borderModifier = PLAYER_BORDER_MODIFIER,
+                accountViewModel = accountViewModel,
+            )
         }
 
         Column(
