+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/podcasts/PodcastEpisodeListItem.kt
@@ -23,16 +23,13 @@ package com.vitorpamplona.amethyst.ui.screen.loggedIn.podcasts
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalContext
@@ -40,23 +37,14 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.service.playback.composable.GetVideoController
-import com.vitorpamplona.amethyst.service.playback.composable.PauseControllerWhenInBackground
-import com.vitorpamplona.amethyst.service.playback.composable.WaveformData
-import com.vitorpamplona.amethyst.service.playback.composable.mediaitem.GetMediaItem
 import com.vitorpamplona.amethyst.ui.navigation.navs.INav
 import com.vitorpamplona.amethyst.ui.navigation.routes.routeFor
 import com.vitorpamplona.amethyst.ui.note.timeAgo
-import com.vitorpamplona.amethyst.ui.note.types.RenderVoicePlayer
+import com.vitorpamplona.amethyst.ui.note.types.PodcastEpisodeAudioPlayer
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.AccountViewModel
 import com.vitorpamplona.amethyst.ui.theme.Size5dp
 import com.vitorpamplona.amethyst.ui.theme.grayText
 import com.vitorpamplona.quartz.nipF4Podcasts.episode.PodcastEpisodeEvent
-
-// NIP-F4 episodes carry no waveform tag, so render a flat baseline. Shared at file scope so
-// every visible row points at the same List<Float> instead of allocating a fresh one per card.
-private const val WAVEFORM_SAMPLES = 96
-private val FLAT_WAVEFORM = WaveformData(List(WAVEFORM_SAMPLES) { 0.4f })
 
 private val PLAYER_SHAPE = Modifier.clip(RoundedCornerShape(12.dp))
 
@@ -91,7 +79,7 @@ fun PodcastEpisodeListItem(
     ) {
         dateStr.takeIf { it.isNotBlank() }?.let {
             Text(
-                text = it.uppercase(),
+                text = it,
                 style = MaterialTheme.typography.labelMedium,
                 color = MaterialTheme.colorScheme.grayText,
             )
@@ -120,39 +108,14 @@ fun PodcastEpisodeListItem(
         }
 
         firstAudio?.let { audio ->
-            val callbackUri = remember(noteEvent) { note.toNostrUri() }
-
-            Row(
-                Modifier.fillMaxWidth().height(72.dp).padding(top = Size5dp),
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                GetMediaItem(
-                    videoUri = audio.url,
-                    title = title,
-                    artworkUri = image,
-                    authorName = note.author?.toBestDisplayName(),
-                    callbackUri = callbackUri,
-                    mimeType = audio.mediaType,
-                    aspectRatio = null,
-                    proxyPort = accountViewModel.httpClientBuilder.proxyPortForVideo(audio.url),
-                    keepPlaying = false,
-                    waveformData = FLAT_WAVEFORM,
-                ) { mediaItem ->
-                    GetVideoController(
-                        mediaItem = mediaItem,
-                        muted = false,
-                    ) { controller ->
-                        PauseControllerWhenInBackground(controller)
-                        RenderVoicePlayer(
-                            mediaItem = mediaItem,
-                            controllerState = controller,
-                            waveform = FLAT_WAVEFORM,
-                            borderModifier = PLAYER_SHAPE,
-                            accountViewModel = accountViewModel,
-                        )
-                    }
-                }
-            }
+            PodcastEpisodeAudioPlayer(
+                audio = audio,
+                note = note,
+                title = title,
+                image = image,
+                borderModifier = PLAYER_SHAPE,
+                accountViewModel = accountViewModel,
+            )
         }
     }
 }
